+++ js/assets/icons/mdi--trash-can-outline.js
@@ -1,0 +1,1 @@
+(()=>{var v=({size:h=24,...t})=>React.createElement("svg",{xmlns:"http://www.w3.org/2000/svg",width:h,height:h,...t},React.createElement("path",{fill:"currentColor",d:"M9 3v1H4v2h1v13a2 2 0 0 0 2 2h10a2 2 0 0 0 2-2V6h1V4h-5V3zM7 6h10v13H7zm2 2v9h2V8zm4 0v9h2V8z"})),a=v;})();