+++ src/assets/icons/mdi--arrow-down-thin.jsx
@@ -1,0 +1,6 @@
+const MdiArrowDownThin = ({ size = 24, ...props }) => (
+    <svg xmlns="http://www.w3.org/2000/svg" width={size} height={size} {...props}>
+        <path fill="currentColor" d="M7.03 13.92h4V5l2.01-.03v8.95h3.99l-5 5Z" />
+    </svg>
+);
+export default MdiArrowDownThin;