+++ src/assets/icons/mdi--arrow-up-thin.jsx
@@ -1,0 +1,6 @@
+const MdiArrowUpThin = ({ size = 24, ...props }) => (
+    <svg xmlns="http://www.w3.org/2000/svg" width={size} height={size} {...props}>
+        <path fill="currentColor" d="M7.03 9.97h4v8.92l2.01.03V9.97h3.99l-5-5Z" />
+    </svg>
+);
+export default MdiArrowUpThin;