+++ src/assets/icons/mdi--plus.jsx
@@ -1,0 +1,6 @@
+const MdiPlus = ({ size = 24, ...props }) => (
+    <svg xmlns="http://www.w3.org/2000/svg" width={size} height={size} {...props}>
+        <path fill="currentColor" d="M19 13h-6v6h-2v-6H5v-2h6V5h2v6h6z" />
+    </svg>
+);
+export default MdiPlus;