+++ src/config.js
@@ -27,13 +27,6 @@ export const paddingProps = {
     'paddingLeft': { type: 'string' }
 };
 
-export const modeProps = {
-    'dimensionMode': { type: 'string' },
-    'minDimensionMode': { type: 'string' },
-    'maxDimensionMode': { type: 'string' },
-    'marginMode': { type: 'string' },
-    'paddingMode': { type: 'string' }
-};
 
 export const flexProps = {
     'flexDirection': { type: 'string' },
@@ -41,8 +34,6 @@ export const flexProps = {
     'justifyContent': { type: 'string' },
     'alignItems': { type: 'string' },
     'alignContent': { type: 'string' },
-    'gapColumn': { type: 'string' },
-    'gapRow': { type: 'string' }
 };
 
 export const flexItemProps = {
@@ -53,6 +44,28 @@ export const flexItemProps = {
     'order': { type: 'number' }
 };
 
+export const gridProps = {
+    'gridTemplateColumns': { type: 'string' },
+    'gridTemplateRows' : { type: 'string' }
+}
+
+export const modeProps = [
+    'dimensionMode',
+    'minDimensionMode',
+    'maxDimensionMode',
+    'marginMode',
+    'paddingMode',
+    'gapMode',
+    'gapInputMode',
+    'gridTemplateColumnsMode',
+    'gridTemplateRowsMode'
+];
+
+export const  sharedProps = {
+    'gap': { type: 'string' },
+}
+
+
 export const COSTERED_LAYOUT_SCHEMA = {
     ...displayProps,
     ...dimensionProps,
@@ -60,13 +73,18 @@ export const COSTERED_LAYOUT_SCHEMA = {
     ...paddingProps,
     ...flexProps,
     ...flexItemProps,
-    ...modeProps
+    ...gridProps,
+    ...sharedProps
 }
 
 export const COSTERED_SCHEMA = Object.assign(
     {},
     ...Object.values(COSTERED_LAYOUT_SCHEMA)
 )
+
+export const MUST_BE_SCOPED = new Set([
+    ...modeProps,
+])
 
 export const BLOCKS_WITH_EDITOR_STYLES = [
     'core/paragraph',
@@ -76,3 +94,42 @@ export const BLOCKS_WITH_EDITOR_STYLES = [
     'core/pullquote',
     'core/table'
 ]
+
+/**
+ * Editor Style Mirror keys.
+ * These three keys are used to determine what block attributes (styles) should be mirrored
+ * in the block editor. The WP Block Editor can be tough to work with as the editor state is
+ * always being updated by something (too many things to list). The Editor Style Mirror (located
+ * in ./src/filters/editor-style-mirror.js) gives us a little bit of control over that state, but
+ * we neeed to specify what styles can be set using these three keys.
+ * To add addtional styles, add the appropriate block attribute (css style in camelCase), make a
+ * group of props like the example below, and include that group in MISC_KEYS.
+ * 
+ * export const customProps = {
+ *   'propertyName': { type: 'string|number' },
+ * };
+ * 
+ */
+
+/** Please don't touch this key. Controls padding styles in the Block Editor. Won't function without this. */
+export const PADDING_KEYS = new Set([
+    {},
+    ...Object.keys(paddingProps)
+]);
+
+/** Please don't touch this key. Controls width/height styles in the Block Editor. Won't function without this. */
+export const SIZE_KEYS = new Set([
+    {},
+   ...Object.keys(dimensionProps)
+]);
+
+/** If you want to extend this, create your own group and include it like the other groups below.  */
+export const MISC_KEYS = new Set([
+    {},
+    ...Object.keys(displayProps),
+    ...Object.keys(flexProps),
+    ...Object.keys(flexItemProps),
+    ...Object.keys(gridProps),
+    ...Object.keys(sharedProps),
+    //...Object.keys(yourGroupGoesHere),
+])